[Update]Bat
박쥐 몬스터 추가.
-> 상태 패턴으로 구현
-> 몬스터의 체력, 이동속도, 공격력은
외부 데이터에서 받아오도록 수정함.

Weapon
-> 몬스터 피격 구현(충돌 확인함, HP 감소 확인함)
문제점: 스크립트가 검에 붙어있는게 아니라 AddComponent로 플레이어에게 붙어있어서 플레이어가 몸통 박치기를 하면 OnTriggerEnter가 호출됨....
 ->해결중

+++ Assets/Manager/SkillManager.cs
@@ -31,10 +31,10 @@ public class SkillManager : Singleton<SkillManager>
 
     private void Init()
     {
-        SkillDictionary.Add(PlayerSkill.Bow, new SkillData(1.0f, 500.0f, 0f, 1));
-        SkillDictionary.Add(PlayerSkill.FireBall, new SkillData(1.0f, 200.0f, 1f, 1));
-        SkillDictionary.Add(PlayerSkill.Bomb, new SkillData(1.0f, 10.0f, 10.0f, 2));
-        SkillDictionary.Add(PlayerSkill.Hook, new SkillData(1.0f, 10.0f, 5.0f, 0));
+        SkillDictionary.Add(PlayerSkill.Bow, new SkillData(1, 500.0f, 0f, 1));
+        SkillDictionary.Add(PlayerSkill.FireBall, new SkillData(1, 200.0f, 1f, 1));
+        SkillDictionary.Add(PlayerSkill.Bomb, new SkillData(1, 10.0f, 10.0f, 2));
+        SkillDictionary.Add(PlayerSkill.Hook, new SkillData(1, 10.0f, 5.0f, 0));
 
         m_skillCount = 4;
     }
@@ -74,11 +74,11 @@ public class SkillManager : Singleton<SkillManager>
 
 public struct SkillData
 {
-    public float m_attackPower { get; }
+    public int m_attackPower { get; }
     public float m_projectileSpeed { get; }
     public float m_projectileRange { get; }
     public int m_cost { get; }
-    public SkillData(float attackPower,float projectileSpeed, float projectileRange, int cost)
+    public SkillData(int attackPower,float projectileSpeed, float projectileRange, int cost)
     {
         m_attackPower = attackPower;
         m_projectileSpeed = projectileSpeed;
