+++ Assets/Manager/WeaponManager.cs
@@ -24,11 +24,11 @@ public class WeaponManager : Singleton<WeaponManager>
 
     private void InitWeaponData()
     {
-        WeaponDataDictionary.Add(PlayerWeapon.Sword, new WeaponData(1.0f, 3.0f, 0.7f,1.0f,new Vector3(0.7f,0.7f,0.7f),new Vector3(1.0f,0.7f,1.0f)));
-        WeaponDataDictionary.Add(PlayerWeapon.Hammer, new WeaponData(1.0f, 3.0f, 0.7f, 1.0f, new Vector3(0.7f, 0.7f, 0.7f), new Vector3(1.0f, 0.7f, 1.0f)));
-        WeaponDataDictionary.Add(PlayerWeapon.Dagger, new WeaponData(1.0f, 3.0f, 0.7f, 1.0f, new Vector3(0.7f, 0.7f, 0.7f), new Vector3(1.0f, 0.7f, 1.0f)));
-        WeaponDataDictionary.Add(PlayerWeapon.GreatSword, new WeaponData(1.0f, 3.0f, 0.7f, 1.0f, new Vector3(0.7f, 0.7f, 0.7f), new Vector3(1.0f, 0.7f, 1.0f)));
-        WeaponDataDictionary.Add(PlayerWeapon.Umbrella, new WeaponData(1.0f, 3.0f, 0.7f, 1.0f, new Vector3(0.7f, 0.7f, 0.7f), new Vector3(1.0f, 0.7f, 1.0f)));
+        WeaponDataDictionary.Add(PlayerWeapon.Sword, new WeaponData(1, 3, 0.7f,1.0f,new Vector3(0.7f,0.7f,0.7f),new Vector3(1.0f,0.7f,1.0f)));
+        WeaponDataDictionary.Add(PlayerWeapon.Hammer, new WeaponData(1, 3, 0.7f, 1.0f, new Vector3(0.7f, 0.7f, 0.7f), new Vector3(1.0f, 0.7f, 1.0f)));
+        WeaponDataDictionary.Add(PlayerWeapon.Dagger, new WeaponData(1, 3, 0.7f, 1.0f, new Vector3(0.7f, 0.7f, 0.7f), new Vector3(1.0f, 0.7f, 1.0f)));
+        WeaponDataDictionary.Add(PlayerWeapon.GreatSword, new WeaponData(1, 3, 0.7f, 1.0f, new Vector3(0.7f, 0.7f, 0.7f), new Vector3(1.0f, 0.7f, 1.0f)));
+        WeaponDataDictionary.Add(PlayerWeapon.Umbrella, new WeaponData(1, 3, 0.7f, 1.0f, new Vector3(0.7f, 0.7f, 0.7f), new Vector3(1.0f, 0.7f, 1.0f)));
     }
 
     private void Start()
@@ -69,13 +69,13 @@ public class WeaponManager : Singleton<WeaponManager>
 
 public struct WeaponData
 {
-    public float m_attackPower { get; }
-    public float m_chargeAttackPower { get; }
+    public int m_attackPower { get; }
+    public int m_chargeAttackPower { get; }
     public float m_normalEffectRange { get; }
     public float m_chargeEffectRange { get; }
     public Vector3 m_normalAttackRange { get; }
     public Vector3 m_chargeAttackRange { get; }
-    public WeaponData(float attackPower, float chargeAttackPower, float normalEffectRange, float chargeEffectRange, Vector3 normalAttackRange, Vector3 chargeAttackRange)
+    public WeaponData(int attackPower, int chargeAttackPower, float normalEffectRange, float chargeEffectRange, Vector3 normalAttackRange, Vector3 chargeAttackRange)
     {
         m_attackPower = attackPower;
         m_chargeAttackPower = chargeAttackPower;
