+++ Assets/Player/Script/PlayerWeaponScript/Weapon.cs
@@ -10,7 +10,7 @@ public abstract class Weapon : MonoBehaviour, IWeapon
     protected PlayerWeaponEffectController m_weaponEffect;
     protected PlayerWeaponController m_weaponController;
     protected WeaponData m_weaponData;
-    protected float m_currentAtk;
+    protected int m_currentAtk;
 
     public abstract void InitWeapon();
     public abstract void UseWeapon(bool isCharge);
